+++ src/types/any.rs
@@ -3,6 +3,8 @@ use crate::conversion::{AsPyPointer, FromPyObject, IntoPy, IntoPyPointer, PyTryF
 use crate::err::{PyDowncastError, PyErr, PyResult};
 use crate::exceptions::PyTypeError;
 use crate::type_object::PyTypeInfo;
+#[cfg(not(PyPy))]
+use crate::types::PySuper;
 use crate::types::{PyDict, PyIterator, PyList, PyString, PyTuple, PyType};
 use crate::{err, ffi, Py, PyNativeType, PyObject, Python};
 use std::cell::UnsafeCell;
@@ -882,6 +884,14 @@ impl PyAny {
     #[inline]
     pub fn py(&self) -> Python<'_> {
         PyNativeType::py(self)
+    }
+
+    /// Return a proxy object that delegates method calls to a parent or sibling class of type.
+    ///
+    /// This is equivalent to the Python expression `super()`
+    #[cfg(not(PyPy))]
+    pub fn py_super(&self) -> PyResult<&PySuper> {
+        PySuper::new(self.get_type(), self)
     }
 }