+++ src/types/mod.rs
@@ -27,6 +27,8 @@ pub use self::mapping::PyMapping;
 pub use self::module::PyModule;
 pub use self::num::PyLong;
 pub use self::num::PyLong as PyInt;
+#[cfg(not(PyPy))]
+pub use self::pysuper::PySuper;
 pub use self::sequence::PySequence;
 pub use self::set::PySet;
 pub use self::slice::{PySlice, PySliceIndices};
@@ -277,6 +279,8 @@ mod list;
 mod mapping;
 mod module;
 mod num;
+#[cfg(not(PyPy))]
+mod pysuper;
 mod sequence;
 mod set;
 mod slice;